Decorate messages and ensure delivery
This adds a DecoratedMessage type which will be sent to all of the
message listeners. This will contain extra information like the type of
the filter that wrote the message and when it was written. It also
contains a `Source` flag which will denote messages that are sent from
the pipeline itself, whose codes are listed in `decorated_messages.go`.

Internally, message channels are now split so each filter has one and
that channel must clear before the filter shuts down. This prevents a
previous race condition where messages may not have been delivered to
the listeners before the pipeline returned.